% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joboffer}
\alias{joboffer}
\title{The text of a job offer, annotated with the package udpipe}
\description{
The text of a job offer, annotated with the package udpipe
}
\examples{
data(joboffer)
str(joboffer)
unique(joboffer$sentence)
}
