% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textrank.R
\name{summary.textrank}
\alias{summary.textrank}
\title{Extract the most important sentences}
\usage{
\method{summary}{textrank}(object, n = 3, keep.sentence.order = FALSE, ...)
}
\arguments{
\item{object}{an object of class textrank}

\item{n}{integer indicating to extract only the top n sentences}

\item{keep.sentence.order}{logical indicating to keep the sentence order in the original \code{data} element of the \code{\link{textrank}} function
or to order it by the pagerank score. Defaults to FALSE - order by pagerank score}

\item{...}{not used}
}
\value{
a character vector with the top \code{n} most important sentences
which were identified by \code{\link{textrank}}
}
\description{
Extract the most important sentences which were identified by \code{\link{textrank}}
}
\seealso{
\code{\link{textrank}}
}
