% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textrank.R
\name{summary.textrank_sentences}
\alias{summary.textrank_sentences}
\title{Extract the most important sentences which were identified with textrank_sentences}
\usage{
\method{summary}{textrank_sentences}(object, n = 3,
  keep.sentence.order = FALSE, ...)
}
\arguments{
\item{object}{an object of class textrank_sentences}

\item{n}{integer indicating to extract only the top n sentences}

\item{keep.sentence.order}{logical indicating to keep the sentence order as provided
in the original \code{data} argument of the \code{\link{textrank_sentences}} function
or to order it by the pagerank score. Defaults to FALSE indicating to order by pagerank score.}

\item{...}{not used}
}
\value{
a character vector with the top \code{n} most important sentences
which were identified by \code{\link{textrank_sentences}}
}
\description{
Extract the most important sentences which were identified by \code{\link{textrank_sentences}}
}
\seealso{
\code{\link{textrank_sentences}}
}
