% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textrank.R
\name{textrank_jaccard}
\alias{textrank_jaccard}
\title{Calculate the distance between 2 vectors based on the Jaccard distance}
\usage{
textrank_jaccard(termsa, termsb)
}
\arguments{
\item{termsa}{a character vector of words}

\item{termsb}{a character vector of words}
}
\value{
The Jaccard distance distance between the 2 vectors
}
\description{
The jaccard distance computes the percentage of terms in the 2 vectors which are overlapping.
}
\examples{
sentencea <- c("I", "like", "champaign")
sentenceb <- c("I", "prefer", "choco")
textrank_jaccard(termsa = sentencea, termsb = sentenceb)
}
