% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textrank.R
\name{textrank_keywords}
\alias{textrank_keywords}
\title{Textrank - extract relevant keywords}
\usage{
textrank_keywords(x, is_relevant, p = 1/3, ngram_max = 5, sep = "-")
}
\arguments{
\item{x}{a character vector of words}

\item{is_relevant}{a logical vector indicating if the word is relevant or not}

\item{p}{percentage of relevant words to keep}

\item{ngram_max}{maximum ngram}

\item{sep}{separator}
}
\value{
Under construction. Will change soon
}
\description{
Textrank - extract relevant keywords
}
\examples{
data(joboffer)
keywords <- textrank_keywords(joboffer$lemma, joboffer$upos \%in\% c("NOUN", "VERB", "ADJ"))
subset(keywords, ngram > 1)
}
\seealso{
\code{\link{textrank}}
}
